# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bella Cucina Restaurant API",
    "description": "API for Bella Cucina Italian Restaurant with authentication",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://bella-cucina-restaurant-production.up.railway.app",
      "description": "Bella Cucina Live Server"
    }
  ],
  "paths": {
    "/menu": {
      "get": {
        "summary": "Get full menu with prices",
        "operationId": "getFullMenu",
        "description": "Returns the complete restaurant menu. No authentication required.",
        "tags": ["Menu"],
        "responses": {
          "200": {
            "description": "Full menu with categories and items"
          }
        }
      }
    },
   "/offers": {
      "get": {
        "summary": "Get current offers and promotions",
        "operationId": "getCurrentOffers",
        "description": "Returns all currently active promotions. Requires OAuth 2.0 token (get token from POST /oauth/token).",
        "tags": ["Offers"],
        "security": [
          {
            "oauth2Auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of active offers"
          },
          "401": {
            "description": "Invalid or missing OAuth token"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Place a new food order",
        "operationId": "placeOrder",
        "description": "Places a new food order. Requires API Key in x-api-key header.",
        "tags": ["Orders"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer_name", "phone", "items", "order_type"],
                "properties": {
                  "customer_name": {
                    "type": "string",
                    "example": "Ahmed Mohamed"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+201012345678"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "item_id": {
                          "type": "integer",
                          "example": 4
                        },
                        "name": {
                          "type": "string",
                          "example": "Margherita Pizza"
                        },
                        "quantity": {
                          "type": "integer",
                          "example": 2
                        },
                        "special_instructions": {
                          "type": "string",
                          "example": "no onions"
                        }
                      }
                    }
                  },
                  "order_type": {
                    "type": "string",
                    "enum": ["delivery", "pickup"],
                    "example": "delivery"
                  },
                  "address": {
                    "type": "string",
                    "example": "15 Tahrir St, Cairo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed successfully"
          },
          "401": {
            "description": "Invalid or missing API Key"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Track order status",
        "operationId": "getOrderStatus",
        "description": "Returns order details. Requires Bearer Token.",
        "tags": ["Orders"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID (e.g. ORD-2026-001)"
          }
        ],
        "responses": {
          "200": {
            "description": "Order details and status"
          },
          "401": {
            "description": "Invalid or missing Bearer Token"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/reservations": {
      "post": {
        "summary": "Make a table reservation",
        "operationId": "createReservation",
        "description": "Creates a new table reservation. Requires Basic Auth (username: bella, password: cucina123).",
        "tags": ["Reservations"],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer_name", "phone", "date", "time", "guests"],
                "properties": {
                  "customer_name": {
                    "type": "string",
                    "example": "Sara Ali"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+201098765432"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-03-10"
                  },
                  "time": {
                    "type": "string",
                    "example": "19:00"
                  },
                  "guests": {
                    "type": "integer",
                    "example": 4
                  },
                  "special_requests": {
                    "type": "string",
                    "example": "window seat please"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation created"
          },
          "401": {
            "description": "Invalid Basic Auth credentials"
          }
        }
      }
    },
    "/delivery/check": {
      "post": {
        "summary": "Check delivery availability for an area",
        "operationId": "checkDeliveryArea",
        "description": "Checks if delivery is available. No authentication required.",
        "tags": ["Delivery"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["area"],
                "properties": {
                  "area": {
                    "type": "string",
                    "example": "Zamalek"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delivery availability info"
          }
        }
      }
    },
    "/loyalty/{phone}": {
      "get": {
        "summary": "Check loyalty points balance",
        "operationId": "getLoyaltyPoints",
        "description": "Returns loyalty points. Requires JWT token (get token from POST /auth/token).",
        "tags": ["Loyalty"],
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer phone number"
          }
        ],
        "responses": {
          "200": {
            "description": "Loyalty account details"
          },
          "401": {
            "description": "Invalid or missing JWT"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Get JWT token",
        "operationId": "getAuthToken",
        "description": "Get a JWT token for authenticated endpoints. Use username: bella, password: cucina123",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "bella"
                  },
                  "password": {
                    "type": "string",
                    "example": "cucina123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key: bella-cucina-api-key-2026"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer Token: bella-cucina-bearer-token-2026"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Username: bella, Password: cucina123"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from POST /auth/token"
      },

      "oauth2Auth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://bella-cucina-restaurant-production.up.railway.app/oauth/token",
            "scopes": {}
          }
        },
        "description": "Client ID: bella-cucina-client, Client Secret: bella-cucina-secret-2026"
      }
    }
  }
}